+++ packages/ragbits-core/CHANGELOG.md
@@ -2,12 +2,13 @@
 
 ## Unreleased
 
+- Integrate Weaviate vector store (#347)
+
 ## 0.19.1 (2025-05-27)
 
 ## 0.19.0 (2025-05-27)
 
 - rename typevars InputT and OutputT to PromptInputT and PromptOutputT (#569)
-- Integrate Weaviate vector store (#347)
 
 ## 0.18.0 (2025-05-22)